AbstractTrees extension: avoid loss of information/semantic ambiguity
If `AbstractTrees.nodevalue(::Symbolic)` and
`AbstractTrees.nodevalue(::SymbolicParameter)` both return `Symbol`,
it's not possible to tell apart a variable from a parameter
programmatically in the tree produced from an expression by
AbstractTrees.jl.

Just deleting the two methods prevents the loss of information, now
`AbstractTrees.nodevalue` is just the identity function for variables
and for parameters.

Fixes #43